CDAP-1583 refactored inmemory program runners - worker, service
